+++ test/jest/babel.config.js
@@ -1,9 +1,5 @@
+const { babelOptions } = require('@folio/stripes-cli');
+
 module.exports = {
-  presets: [
-    '@babel/preset-env',
-    '@babel/preset-react',
-  ],
-  plugins: [
-    '@babel/plugin-transform-class-properties',
-  ],
+  ...babelOptions,
 };